requirements.txt
ultralytics==8.1.28
opencv-python>=4.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
Pillow>=9.5.0,<11.0.0
